chore: rename spline component
