+++ frontend/src/components/ui/spline/Spline.vue
