+++ frontend/src/components/ui/spline/index.ts
